Move report/agent tests under tests/ and split OLAP vs TPC-C classify rules.
Keep fixtures with each package; TPC-C compare-delta lives in tpcc/classify_rules.py.
