+++ .github/scripts/utils/perfomance_tests_status/duty_agent/.gitignore
@@ -1,4 +1,5 @@
 runs/
 __pycache__/
+tests/__pycache__/
 *.pyc
 .token_cache
