+++ .github/scripts/utils/perfomance_tests_status/olap/classify_rules.py
@@ -1,4 +1,4 @@
-"""Pure OLAP Now classify / compare rules (shared by generate.py + unit tests)."""
+"""Pure OLAP Now classify / compare rules (generate.py + unit tests). TPC-C lives in tpcc/."""
 
 from __future__ import annotations
 
@@ -169,45 +169,4 @@ def compare_delta_olap(prev: dict, now: dict) -> str:
         return "worse"
     if fail_down or slow_down or nb < pb:
         return "better"
-    return "same"
-
-
-def tpcc_hard_band(side: dict | None) -> int:
-    if not side:
-        return 0
-    st = side.get("status") or ""
-    if st == "broken" or (side.get("n_broken") or 0) > 0:
-        return 2
-    if st in ("lat", "tpmc", "both", "regression") or (side.get("n_lat") or 0) or (side.get("n_tpmc") or 0):
-        return 1
-    return 0
-
-
-def compare_delta_tpcc(prev: dict, now: dict, lat_tol: float = 10.0, tpmc_tol: float = 10.0) -> str:
-    pb = tpcc_hard_band(prev)
-    nb = tpcc_hard_band(now)
-    lat_step = 0.0
-    tpmc_step = 0.0
-    if prev.get("lat_pct") is not None and now.get("lat_pct") is not None:
-        lat_step = float(now["lat_pct"]) - float(prev["lat_pct"])
-    if prev.get("tpmc_pct") is not None and now.get("tpmc_pct") is not None:
-        # positive step = tpmC got worse (lower)
-        tpmc_step = float(prev["tpmc_pct"]) - float(now["tpmc_pct"])
-    broken_up = (now.get("n_broken") or 0) > (prev.get("n_broken") or 0)
-    broken_down = (now.get("n_broken") or 0) < (prev.get("n_broken") or 0)
-    if nb > pb or broken_up:
-        return "worse-hot" if pb >= 1 else "worse"
-    if nb < pb or broken_down:
-        return "better"
-    if pb >= 1 and nb >= 1:
-        lat_worse = lat_step >= lat_tol
-        lat_better = lat_step <= -lat_tol
-        tpmc_worse = tpmc_step >= tpmc_tol
-        tpmc_better = tpmc_step <= -tpmc_tol
-        if (lat_worse and tpmc_better) or (lat_better and tpmc_worse):
-            return "mixed"
-        if lat_worse or tpmc_worse:
-            return "worse-hot"
-        if lat_better or tpmc_better:
-            return "better"
     return "same"
