+++ .github/scripts/utils/perfomance_tests_status/tpcc/.gitignore
@@ -1,0 +1,4 @@
+out/
+__pycache__/
+tests/__pycache__/
+*.pyc
